+++ src/InputPassword/InputPassword.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import Icon from "../Icon";
 import Styles from "./inputPassword.style";
 import InputWrapper from "../InputWrapper";
-import classNames from "../../utils/classNames";
+import classNames from "../_utils/classNames";
 
 export default function InputPassword({
   id,
